+++ packages/server/shared/src/index.ts
@@ -22,7 +22,7 @@ export * from './lib/promise-handler';
 export * from './lib/security/encryption';
 export * from './lib/security/encryption-key-initializer';
 export * from './lib/security/local-store';
-export * from './lib/security/object-transformer';
+export * from './lib/security/encrypt-compress';
 export * from './lib/semaphore';
 export * from './lib/stop-handlers';
 export * from './lib/system';
